+++ egret-webpack-bundler/test/simple-project/src/c.ts
@@ -1,0 +1,15 @@
+module aa {
+
+
+    export class C extends A {
+
+        // static xxx = Main1.a;
+
+        constructor() {
+            console.log('hello')
+            super();
+        }
+
+    }
+
+}